requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.1
Flask-Marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.0.0
python-dotenv==1.0.1
referencing==0.33.0
rpds-py==0.17.1
SQLAlchemy==2.0.27
tomli==2.0.1
Werkzeug==3.0.1
PyYAML==6.0.1
